requirements.txt
numpy
pillow
torch
torchvision
pandas
matplotlib
seaborn
lightning
